Fix articles without attributes not displaying correctly.
Articles that do not have an entry in the articles attributes table do
not get read in by Shopware as articles and therefor would not appear in
the articles list in the backend, nor would they appear in the frontend.
At least a default entry must have been generated for that article.

The old code checked whether the generated SQL part for the attributes
is empty and, if so, would skip the INSERT statement for it. This is an
error in the logic and has been removed. A wrong extra value articleID
that would cause an error in the generated SQL has also been removed.

The UPDATE statement for the attributes also had an error where it would
always prepend the list of values to update with a comma which generates
an SQL syntax error.